Former-commit-id: bd9316f4b9796b0d4f24bfc46db3cbf809e5c13f
Former-commit-id: 832994c12c7b4e4678f1bf1850cd618cc43bcb71

+++ registration-processor/stages/packet-receiver-stage/src/main/java/io/mosip/registration/processor/packet/receiver/stage/PacketReceiverStage.java
@@ -79,7 +79,7 @@ public class PacketReceiverStage extends MosipVerticleAPIManager {
 	 * deploys this verticle.
 	 */
 	public void deployVerticle() {
-		this.mosipEventBus = this.getEventBus(this, clusterManagerUrl);
+		this.mosipEventBus = this.getEventBus(this, clusterManagerUrl, 50);
 
 	}
 
@@ -105,9 +105,9 @@ public class PacketReceiverStage extends MosipVerticleAPIManager {
 	 */
 	private void routes(Router router) {
 		
-		router.post("/packetreceiver/registration-processor/registrationpackets/v1.0").handler(ctx -> {
+		router.post("/packetreceiver/registration-processor/registrationpackets/v1.0").blockingHandler(ctx -> {
 			processURL(ctx);
-		}).failureHandler(failureHandler -> {
+		}, false).failureHandler(failureHandler -> {
 			this.setResponse(failureHandler, globalExceptionHandler.handler(failureHandler.failure()),APPLICATION_JSON);
 		});
 		
@@ -146,8 +146,11 @@ public class PacketReceiverStage extends MosipVerticleAPIManager {
 					"", e.getMessage() + ExceptionUtils.getStackTrace(e));
 			throw new UnexpectedException(e.getMessage());
 		} finally {
-			if (file.exists())
-				deleteFile(file);
+			if(file != null) {
+				if (file.exists()) {
+					deleteFile(file);
+				}
+			}
 		}
 	}
